alarino_backend/pip-requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.6.1
iniconfig==1.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pip==24.2
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytest==8.3.4
python-dotenv==1.1.0
setuptools==75.8.0
six==1.16.0
SQLAlchemy==2.0.40
typing_extensions==4.13.0
Werkzeug==3.1.3
wheel==0.44.0
zipp==3.21.0
